TEST: Add integration tests for node package
This adds a new integration test that builds the node package, imports
it from a node script, and runs all of the "official" test cases from
the `tests/data/tests.json` file there in node.

I've also refactored the main function to be called `apply` to match the
original implementation. I'll be making this same refactor shortly in
the main and python implementations.

I also renamed the `javascript` compilation feature to `wasm`.