FIX: Resolve distribution issues (#16)
* FIX: update repo name

I guess the `repository` reference on the GH context includes the owner
name, so this updates it to just set it manually. It's unlikely to
change ever.

* FIX: only try to upload manylinux wheels

Non manylinux wheels aren't supported, and I forgot they'd be included
in the artifact.

* REF: require initial build for all dist steps

This makes it so we won't distribute the wheels if any step in the main
distribution job (for the cargo, npm, and python sdists) fails.