FIX: persist cargo version check as artifact
Now that we're gating the python wheel deploy on the "main" deploy step,
the cargo check was returning that it's not a new version (since we
publish the crate in the first step).

This persists the version check result as an artifact, and then checks
it in the wheel distribution steps.